# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storyboard/generate-image:
    post:
      operationId: generateSceneImage
      tags: [storyboard]
      summary: Generate an AI image for a storyboard scene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageGenerationRequest"
      responses:
        "200":
          description: Image generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerationResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    CharacterProfile:
      type: object
      properties:
        characterId:
          type: string
          description: English lowercase hyphenated slug, unique across all characters
        name:
          type: string
          description: Character name in the output language
        species:
          type: string
          description: Species or archetype in the output language
        appearance:
          type: string
          description: 3-4 sentences describing the character visually
        clothing:
          type: string
          description: Garments, accessories, footwear in the output language. "—" if none.
        personality:
          type: string
          description: 2-3 sentences covering motivation, baseline, and behavioral tell
        distinctiveFeatures:
          type: string
          description: 2-4 precise, artist-reproducible visual markers
        voiceStyle:
          type: string
          description: "English only. One sentence: pitch, tempo, texture, accent, delivery style."
      required:
        - characterId
        - name
        - species
        - appearance
        - personality
        - distinctiveFeatures

    # ─── Story Layer Schemas ───────────────────────────────────────────────────
...
    ImageProvider:
      type: string
      enum: [stability, openai, replicate, fal, pollinations]
      description: AI image generation provider
    ImageGenerationRequest:
      type: object
      description: Request to generate an image for a scene
      properties:
        storyboardId:
          type: string
          description: Storyboard identifier for asset persistence
        sceneNumber:
          type: integer
          description: Scene number to generate image for
        sceneImagePrompt:
          type: string
          description: Main image prompt from SceneImagePrompt.sceneImagePrompt
        provider:
          $ref: "#/components/schemas/ImageProvider"
        characterProfiles:
          type: array
          items:
            $ref: "#/components/schemas/CharacterProfile"
          description: Character profiles for reference locking
        characterVisualContinuity:
          type: string
          description: Per-character visual continuity state from SceneImagePrompt
        colorPalette:
          type: string
          description: Scene color palette
        cinematicMood:
          type: string
          description: Cinematic mood compound phrase
        renderStyle:
          type: string
          description: Render style specification
        visualEngine:
          type: string
          description: Visual engine applied (PresentEngine, FlashbackEngine, etc.)
      required:
        - sceneNumber
        - sceneImagePrompt
    ImageGenerationResult:
      type: object
      description: Result of an image generation request
      properties:
        imageStatus:
          type: string
          enum: [success, error]
          description: Generation status
        imageUrl:
          type: string
          description: URL of the generated image (present when imageStatus is success)
        imageProvider:
          $ref: "#/components/schemas/ImageProvider"
        generationTime:
          type: number
          description: Time taken to generate the image in seconds
        generationError:
          type: string
          description: Error message (present when imageStatus is error)
      required:
        - imageStatus
        - imageProvider
        - generationTime

    # ─── Asset Persistence ──────────────────────────────────────────────────────────
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error