# lib/api-zod/src/generated/types/imageGenerationRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterProfile } from './characterProfile';
import type { ImageProvider } from './imageProvider';

/**
 * Request to generate an image for a scene
 */
export interface ImageGenerationRequest {
  /** Storyboard identifier for asset persistence */
  storyboardId?: string;
  /** Scene number to generate image for */
  sceneNumber: number;
  /** Main image prompt from SceneImagePrompt.sceneImagePrompt */
  sceneImagePrompt: string;
  provider?: ImageProvider;
  /** Character profiles for reference locking */
  characterProfiles?: CharacterProfile[];
  /** Per-character visual continuity state from SceneImagePrompt */
  characterVisualContinuity?: string;
  /** Scene color palette */
  colorPalette?: string;
  /** Cinematic mood compound phrase */
  cinematicMood?: string;
  /** Render style specification */
  renderStyle?: string;
  /** Visual engine applied (PresentEngine, FlashbackEngine, etc.) */
  visualEngine?: string;
}